requirements.txt
fastapi
uvicorn[standard]
gunicorn
langchain-google-genai
langchain-core
pandas
sqlite3
python-dotenv
pydantic
httpx
python-multipart
structlog
Werkzeug